requirements.txt
python-dotenv==0.17.1
flask==2.0.3
requests==2.27.1
gunicorn==20.1.0
Werkzeug~=2.0.3